_src/requirements.txt
beautifulsoup4
discord.py
mako
Pillow
pyTelegramBotAPI
python-dotenv
pyyaml
shortuuid
requests
